+++ tests/Volcanus/FileUploader/Test/File/SymfonyFileTest.php
@@ -328,7 +328,7 @@ class SymfonyFileTest extends TestCase
             new \RecursiveDirectoryIterator($this->tempDir)
         );
         foreach ($it as $file) {
-            if ($file->isFile() && $file->getBaseName() !== '.gitignore') {
+            if ($file->isFile() && $file->getBaseName() !== '.gitignore' && $file->getBaseName() !== '.gitkeep') {
                 unlink($file);
             }
         }
